fix(_wikilist): items and fullitems could have members flipped
For definitions lists, the LIST_PATTERN_FORMAT regex matches two
fullitems and two items. Only the order of those are flipped from
each other. This is because the first item in the regex is the
first match("item"), and the second is the second. For the
fullitem however, it is the other way around, as everything is
part of the first fullitem.

Fixes #72